server/hackerator/requirements.txt
gunicorn
Flask
Flask-Bootstrap4
Flask-SocketIO
